[WCNCR00379708] nic: Update function return value for scan
[Description]
1. Changed the return type of several scan-related functions
   from uint32_t to uint16_t.

[Test]
Pass

MTK-Commit-Id: 0fd48c40efc2393682d6ba79b27c7058bc1ffa5a

CR-Id: {WCNCR00379708,WCNCR00379709,WCNCR00379710,WCNCR00379711,WCNCR00379712}
Feature: scan
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/9907402
Test: srv_pf_nep_sanity
ODB-Check: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Commit-Check: srv_check_service <srv_check_service@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
GAI-Code-Review: srv_ai_review001
AutoUT-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>

+++ include/nic_uni_cmd_event.h
@@ -2238,7 +2238,7 @@ struct UNI_CMD_SCAN {
 	uint8_t aucTlvBuffer[];
 } __KAL_ATTRIB_PACKED__;
 
-typedef uint32_t(*PFN_UNI_CMD_SCAN_TAG_HANDLER) (struct ADAPTER
+typedef uint16_t(*PFN_UNI_CMD_SCAN_TAG_HANDLER) (struct ADAPTER
 	*ad, uint8_t *buf, struct CMD_SCAN_REQ_V2 *cmd);
 
 struct UNI_CMD_SCAN_TAG_HANDLE {
@@ -2246,7 +2246,7 @@ struct UNI_CMD_SCAN_TAG_HANDLE {
 	PFN_UNI_CMD_SCAN_TAG_HANDLER pfHandler;
 };
 
-typedef uint32_t(*PFN_UNI_CMD_SCHED_SCAN_TAG_HANDLER) (struct ADAPTER
+typedef uint16_t(*PFN_UNI_CMD_SCHED_SCAN_TAG_HANDLER) (struct ADAPTER
 	*ad, uint8_t *buf, struct CMD_SCHED_SCAN_REQ *cmd);
 
 struct UNI_CMD_SCHED_SCAN_TAG_HANDLE {
